+++ semver-gradle-plugin/main/kotlin/internal/SemVerException.kt
@@ -1,0 +1,3 @@
+package com.javiersc.semver.gradle.plugin.internal
+
+internal class SemVerException(override val message: String) : Exception(message)